ugh, cast to signed int so arithmetic is right
h8 u, integers

+++ apps/msp430-accel/accel.c
@@ -98,7 +98,7 @@ int main (void) {
             continue;
         }
 
-        avg = avg + ((cur_reading - avg) / NUM_AVG);
+        avg = avg + (((signed int)cur_reading - avg) / NUM_AVG);
     }
 
     asm volatile ("MOV.W %0, R12" ::"m"(min));
